Fix types for `ExposureConfig` struct fields

+++ diag/diag.go
@@ -43,15 +43,15 @@ type DiagnosisKey struct {
 // ExposureConfig represents the parameters for detecting exposure.
 // @see https://developer.apple.com/documentation/exposurenotification/enexposureconfiguration
 type ExposureConfig struct {
-	MinimumRiskScore                 int   `json:"minimumRiskScore"`
-	AttenuationLevelValues           []int `json:"attenuationLevelValues"`
-	AttenuationWeight                int   `json:"attenuationWeight"`
-	DaysSinceLastExposureLevelValues []int `json:"daysSinceLastExposureLevelValues"`
-	DaysSinceLastExposureWeight      int   `json:"daysSinceLastExposureWeight"`
-	DurationLevelValues              []int `json:"durationLevelValues"`
-	DurationWeight                   int   `json:"durationWeight"`
-	TransmissionRiskLevelValues      []int `json:"transmissionRiskLevelValues"`
-	TransmissionRiskWeight           int   `json:"transmissionRiskWeight"`
+	MinimumRiskScore                 uint8   `json:"minimumRiskScore"`
+	AttenuationLevelValues           []int   `json:"attenuationLevelValues"`
+	AttenuationWeight                float32 `json:"attenuationWeight"`
+	DaysSinceLastExposureLevelValues []int   `json:"daysSinceLastExposureLevelValues"`
+	DaysSinceLastExposureWeight      float32 `json:"daysSinceLastExposureWeight"`
+	DurationLevelValues              []int   `json:"durationLevelValues"`
+	DurationWeight                   float32 `json:"durationWeight"`
+	TransmissionRiskLevelValues      []int   `json:"transmissionRiskLevelValues"`
+	TransmissionRiskWeight           float32 `json:"transmissionRiskWeight"`
 }
 
 // Repository defines an interface for storing and retrieving diagnosis keys
